# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Outreach Service API
  version: "1.0.0"
  description: API per gestione outreach email aziende
servers:
  - url: https://outreach.example.com
    description: Produzione
  - url: http://localhost:3002
    description: Sviluppo
paths:
...
  /api/users:
    post:
      summary: Crea utente (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 100 }
                surname: { type: string, maxLength: 100 }
                email: { type: string, format: email }
                role: { type: string, enum: [superadmin, admin, collaboratore] }
              required: [name, email]
      responses:
        "200": { description: Creato }
        "409": { description: Email già registrata }
  /api/users/{id}:
    patch:
      summary: Aggiorna ruolo/status utente (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [superadmin, admin, collaboratore] }
                status: { type: string, enum: [active, disabled] }
      responses:
        "200": { description: Aggiornato }
...
  /admin/audit:
    get:
      summary: Log audit con filtri e paginazione (admin only)
      parameters:
        - { name: action, in: query, schema: { type: string } }
        - { name: days, in: query, schema: { type: integer } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
      responses:
        "200":
          description: Pagina HTML audit log
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
security:
  - cookieAuth: []